add exercises

+++ Ejercicios/src/ejercicio1/Calculadora.java
@@ -4,19 +4,19 @@ import javax.print.DocFlavor;
 
 public class Calculadora {
 
-    public Long sumar(Long num1, Long num2) {
+    public double sumar(double num1, double num2) {
         return num1 + num2;
     }
 
-    public Long restar(Long num1, Long num2) {
+    public double restar(double num1, double num2) {
         return num1 - num2;
     }
 
-    public Long multiplicar(Long num1, Long num2) {
+    public double multiplicar(double num1, double num2) {
         return num1 * num2;
     }
 
-    public Long dividir(Long num1, Long num2) {
+    public double dividir(double num1, double num2) {
         return num1 / num2;
     }
 }
